Add dir App component
