+++ Eremenko/src/normalise.css
@@ -1,48 +1,3 @@
-@font-face {
-    font-family: 'Roboto';
-    src: url('fonts/roboto-regular.eot');
-    src: url('fonts/roboto-regular.eot?#iefix') format('embedded-opentype'),
-    url('fonts/roboto-regular.woff2') format('woff2'),
-    url('fonts/roboto-regular.woff') format('woff'),
-    url('fonts/roboto-regular.ttf') format('truetype');
-    font-weight: 400;
-    font-style: normal;
-}
-
-html,
-body {
-    height: 100%;
-    margin: 0;
-    padding: 0;
-    font-family: 'Roboto', sans-serif;
-}
-
-.wrapper {
-    display: -webkit-box;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -ms-flex-direction: column;
-    flex-direction: column;
-    height: 100%;
-    margin: 0 auto;
-}
-
-.container {
-    -webkit-box-flex: 1;
-    -ms-flex: 1 0 auto;
-    flex: 1 0 auto;
-    width: 100%;
-    margin: 0 auto;
-}
-
-.footer {
-    -webkit-box-flex: 0;
-    -ms-flex: 0 0 auto;
-    flex: 0 0 auto;
-}
-
 article,
 aside,
 footer,
