+++ Eremenko/src/vendors.js
@@ -1,0 +1,2 @@
+import './normalise.css';
+import 'bootstrap/dist/css/bootstrap.min.css';
